[ADD] ES Sync

+++ internal/utils/genNum.go
@@ -5,8 +5,9 @@ import (
 	"time"
 )
 
-func GenFloatNum(min, max int) float64 {
+func GenFloatNum(min, max int) float32 {
+
 	rand.Seed(time.Now().UnixNano())
 	randNum := rand.Intn(max-min) + min
-	return float64(randNum)
+	return float32(randNum)
 }
